+++ mvlab7/app/src/main/java/com/example/myapplication/MonthlyReceipts.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 public class MonthlyReceipts {
     public Map<String, Map<String, Receipt> > detail;
-    public Map<String, Integer> summary;
+    public Map<String, Double> summary;
 
     public MonthlyReceipts() {}
 }
